requirements.txt
telebot~=0.0.4
Pillow~=9.5.0